More innovative than washi?
These existing items and projects have become industrial paper, office supplies stickers and wall-mounted wallpaper in Japan and are posted on Amazon even, I give an example of the subject because this production material is a product that can be called an ancient craft, I myself chose the preference of articles from years 76 and 92. Come on, are there other sources to develop the pulp business and become other innovative materials? Apakah ini dapat dikembangkan lagi? Can this be developed again? for new projects later, looking for new investors....
from the source Awake 1976 and Awake 1992